KB-9/US-1.1: Implement Brand CRUD with Admin security and MapStruct
- Create Brand entity and repository
- Implement BrandService with business logic (duplicate check, UUID lookup)
- Add BrandController with @PreAuthorize for Admin roles
- Configure MapStruct for DTO mapping
- Standardize error responses (404, 400, 401, 403)

+++ Backend/src/main/java/com/krilicar/utils/JwtUtils.java
@@ -20,7 +20,7 @@ public class JwtUtils {
     private String jwtSecret;
 
     @Value("${jwt.expiration.ms}")
-    private int jwtExpirationMs;
+    private long jwtExpirationMs;
 
     // Clé de signature JWT
     private Key key() {
